@@ -24,43 +26,46 @@
 	"main": "./dist/electron/main.js",
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^8.2.10",
-		"@angular/common": "^8.2.10",
-		"@angular/compiler": "^8.2.10",
-		"@angular/core": "^8.2.10",
-		"@angular/forms": "^8.2.10",
-		"@angular/platform-browser": "^8.2.10",
-		"@angular/platform-browser-dynamic": "^8.2.10",
-		"@angular/router": "^8.2.10",
+		"@angular/animations": "^8.2.14",
+		"@angular/common": "^8.2.14",
+		"@angular/compiler": "^8.2.14",
+		"@angular/core": "^8.2.14",
+		"@angular/forms": "^8.2.14",
+		"@angular/platform-browser": "^8.2.14",
+		"@angular/platform-browser-dynamic": "^8.2.14",
+		"@angular/router": "^8.2.14",
+		"@logigator/logigator-shared-comps": "latest",
 		"@ngx-translate/core": "^11.0.1",
 		"@ngx-translate/http-loader": "^4.0.0",
-		"bootstrap": "^4.3.1",
+		"bootstrap": "^4.4.1",
+		"express": "^4.17.1",
 		"file-saver": "^2.0.2",
 		"ngx-electron": "^2.1.1",
-		"ngx-toastr": "^11.1.0",
+		"ngx-toastr": "^11.2.1",
 		"node-fetch": "^2.6.0",
-		"pixi.js": "^5.1.5",
+		"pixi.js": "^5.2.0",
 		"run-script-os": "^1.0.7",
 		"rxjs": "~6.4.0",
 		"tslib": "^1.10.0",
 		"zone.js": "~0.9.1"
 	},
 	"devDependencies": {
-		"@angular-builders/custom-webpack": "^8.2.0",
-		"@angular-devkit/build-angular": "^0.803.9",
-		"@angular/cli": "^8.3.9",
-		"@angular/compiler-cli": "^8.2.10",
-		"@angular/language-service": "^8.2.10",
-		"@ngtools/webpack": "^8.3.8",
+		"@angular-builders/custom-webpack": "^8.4.1",
+		"@angular-devkit/build-angular": "^0.803.20",
+		"@angular/cli": "^8.3.20",
+		"@angular/compiler-cli": "^8.2.14",
+		"@angular/language-service": "^8.2.14",
+		"@ngtools/webpack": "^8.3.20",
+		"@types/express": "^4.17.2",
 		"@types/file-saver": "^2.0.1",
 		"@types/gtag.js": "0.0.3",
 		"@types/jasmine": "~3.3.8",
 		"@types/jasminewd2": "^2.0.8",
 		"@types/node": "~8.9.4",
-		"@types/node-fetch": "^2.5.2",
-		"codelyzer": "^5.0.0",
-		"electron": "^6.0.12",
-		"electron-packager": "^14.0.6",
+		"@types/node-fetch": "^2.5.4",
+		"codelyzer": "^5.2.0",
+		"electron": "^6.1.7",
+		"electron-packager": "^14.1.1",
 		"electron-reload": "^1.5.0",
 		"env-cmd": "^10.0.1",
 		"html-loader": "^0.5.5",
