@@ -19,6 +19,7 @@ export class AuthenticationHandler {
 		ipcMain.on('logingoogle', () => this.onGoogleLogin());
 		ipcMain.on('logintwitter', () => this.onTwitterLogin());
 		ipcMain.on('loginemail', (event, args) => this.onEmailLogin(args));
+		ipcMain.on('registeremail', (event, args) => this.onEmailRegister(args));
 		ipcMain.on('logout', () => this.onLogout());
 	}
 
@@ -93,7 +94,32 @@ export class AuthenticationHandler {
 		}
 	}
 
-	private async onEmailLogin(args: {email: string, password: string}) {
+	private async onEmailLogin(args: {user: string, password: string}) {
+		const resp = await fetch(getApiUrl() + '/auth/login-email', {
+			method: 'post',
+			body: JSON.stringify(args),
+			headers: { 'Content-Type': 'application/json' }
+		});
+		if (!resp.ok) {
+			this.sendLoginResponse(false, 'email', await resp.json());
+			return;
+		}
+		await this.setLoggedIn(resp.headers.raw()['set-cookie'].find(c => c.includes('auth-token')));
+		this.sendLoginResponse(true, 'email');
+	}
+
+	private async onEmailRegister(args: {username: string, email: string, password: string, recaptcha: string}) {
+		const resp = await fetch(getApiUrl() + '/auth/register-email', {
+			method: 'post',
+			body: JSON.stringify(args),
+			headers: { 'Content-Type': 'application/json' }
+		});
+		if (!resp.ok) {
+			this.sendLoginResponse(false, 'email', await resp.json());
+			return;
+		}
+		await this.setLoggedIn(resp.headers.raw()['set-cookie'].find(c => c.includes('auth-token')));
+		this.sendLoginResponse(true, 'email');
 	}
 
 	private onLogout() {
@@ -147,7 +173,7 @@ export class AuthenticationHandler {
 			headers: { 'Content-Type': 'application/json' }
 		});
 		if (!resp.ok) {
-			this.sendLoginResponse(false, 'google');
+			this.sendLoginResponse(false, 'google', await resp.json());
 			return;
 		}
 		await this.setLoggedIn(resp.headers.raw()['set-cookie'].find(c => c.includes('auth-token')));
@@ -161,7 +187,7 @@ export class AuthenticationHandler {
 			headers: { 'Content-Type': 'application/json' }
 		});
 		if (!resp.ok) {
-			this.sendLoginResponse(false, 'twitter');
+			this.sendLoginResponse(false, 'twitter', await resp.json());
 			return;
 		}
 		await this.setLoggedIn(resp.headers.raw()['set-cookie'].find(c => c.includes('auth-token')));
