@@ -7,7 +7,8 @@ import {fromEvent, Subject} from 'rxjs';
 import {DOCUMENT} from '@angular/common';
 import {takeUntil} from 'rxjs/operators';
 import {ProjectsService} from './services/projects/projects.service';
-import {TranslateService} from '@ngx-translate/core';
+import {LangChangeEvent, TranslateService} from '@ngx-translate/core';
+import {ElementProviderService} from './services/element-provider/element-provider.service';
 
 @Component({
 	selector: 'app-root',
@@ -30,7 +31,8 @@ export class AppComponent implements OnInit, OnDestroy {
 		private shortcuts: ShortcutsService,
 		private projects: ProjectsService,
 		@Inject(DOCUMENT) private document: HTMLDocument,
-		private translate: TranslateService
+		private translate: TranslateService,
+		private elementProviderService: ElementProviderService
 	) {
 		this.setGoogleAnalytics();
 		this.initTranslation();
@@ -62,23 +64,27 @@ export class AppComponent implements OnInit, OnDestroy {
 	}
 
 	public get showSettingsInfoBox(): boolean {
-		return this.workMode.currentWorkMode === 'buildComponent' ||  (this.projects.currProject && this.selection.isSingleSelect());
+		const seElTypeId = this.selectedElemTypeId;
+		return seElTypeId === undefined ? false : this.elementProviderService.getElementById(seElTypeId).showSettings;
 	}
 
-	public get selectionMode(): 'type' | 'placed' {
+	public get selectedElemTypeId(): number {
 		if (this.workMode.currentWorkMode === 'buildComponent') {
-			return 'type';
+			return this.workMode.currentComponentToBuild;
+		} else {
+			const selectedIds = this.selection.selectedIds();
+			if (!selectedIds || selectedIds.length === 0 || selectedIds.length > 1) {
+				return undefined;
+			}
+			const elemType = this.projects.currProject.currState.getElementById(this.selection.selectedIds()[0]);
+			if (!elemType) return undefined;
+			return elemType.typeId;
 		}
-		return 'placed';
-	}
-
-	public get selectedCompTypeId(): number {
-		return this.workMode.currentComponentToBuild;
 	}
 
 	public get selectedCompId(): number {
-		if (!this.selection.selectedIds()) {
-			return null;
+		if (!this.selection.selectedIds() || this.workMode.currentWorkMode === 'buildComponent') {
+			return undefined;
 		}
 		return this.selection.selectedIds()[0];
 	}
@@ -107,8 +113,17 @@ export class AppComponent implements OnInit, OnDestroy {
 
 	private initTranslation() {
 		this.translate.addLangs(['en', 'de']);
-		this.translate.setDefaultLang('en');
-		this.translate.use('en');
+		const lang = window.localStorage.getItem('lang');
+		if (lang) {
+			this.translate.setDefaultLang(lang);
+			this.translate.use(lang);
+		} else {
+			this.translate.setDefaultLang('en');
+			this.translate.use('en');
+		}
+		this.translate.onLangChange.subscribe((e: LangChangeEvent) => {
+			window.localStorage.setItem('lang', e.lang);
+		});
 	}
 
 	private setGoogleAnalytics() {
