@@ -5,6 +5,7 @@ import {ProjectInteractionService} from '../../../services/project-interaction/p
 import {InteractionAction} from '../../../models/interaction-action';
 import {UserService} from '../../../services/user/user.service';
 import {ProjectsService} from '../../../services/projects/projects.service';
+import {ImageExportService} from '../../../services/image-export/image-export.service';
 
 @Component({
 	selector: 'app-file-dropdown',
@@ -21,7 +22,8 @@ export class FileDropdownComponent implements OnInit {
 		private projectSaveService: ProjectSaveManagementService,
 		private projectInteraction: ProjectInteractionService,
 		private user: UserService,
-		private projects: ProjectsService
+		private projects: ProjectsService,
+		private imageExportService: ImageExportService
 	) { }
 
 	ngOnInit() {
@@ -82,7 +84,8 @@ export class FileDropdownComponent implements OnInit {
 		this.close();
 	}
 
-	public screenshot() {
+	public screenshot(type: 'jpeg' | 'png' | 'svg') {
+		this.imageExportService.exportImage(type);
 		this.close();
 	}
 }
