@@ -1,15 +1,30 @@
 import {ElementType} from '../element-type';
 import {environment} from '../../../../environments/environment';
+import {ElementTypeId} from '../element-type-ids';
 
 export const not: ElementType = {
+	id: ElementTypeId.NOT,
+
 	name: 'ELEMENT_TYPE.BASIC.NOT.NAME',
-	numInputs: 1,
+
+	category: 'basic',
+
+	symbol: '!',
+
+	showSettings: true,
+	showSettingsForType: true,
+	showInConstructionBox: true,
+
+	description: 'ELEMENT_TYPE.BASIC.NOT.DESCRIPTION',
+
+	isRotatable: true,
+	rotation: 0,
+
 	numOutputs: 1,
+
+	numInputs: 1,
 	minInputs: 1,
 	maxInputs: 1,
-	width: environment.componentWidth,
-	symbol: '1',
-	description: 'ELEMENT_TYPE.BASIC.NOT.DESCRIPTION',
-	rotation: 0,
-	category: 'basic'
+
+	width: environment.componentWidth
 };
