@@ -1,14 +1,28 @@
 import {ElementType} from '../element-type';
+import {ElementTypeId} from '../element-type-ids';
 
 export const wire: ElementType = {
+	id: ElementTypeId.WIRE,
+
 	name: 'wire',
-	numInputs: 0,
+
+	category: 'basic',
+
+	symbol: '',
+
+	showSettings: false,
+	showSettingsForType: false,
+	showInConstructionBox: false,
+
+	description: '',
+
+	rotation: 0,
+
 	numOutputs: 0,
+
+	numInputs: 0,
 	minInputs: 0,
 	maxInputs: 0,
+
 	width: 0,
-	symbol: '',
-	description: '',
-	rotation: 0,
-	category: 'basic'
 };
