@@ -1,16 +1,52 @@
+import {ProjectsService} from '../../services/projects/projects.service';
+import {Element} from '../element';
+
 export interface ElementType {
+	id: number;
+
 	// translation id for predefined comps
 	name: string;
-	numInputs: number;
-	numOutputs: number;
-	maxInputs: number;
-	minInputs: number;
-	width: number;
+
+	category: 'basic' | 'advanced' | 'plug' | 'io' | 'user';
+
 	symbol: string;
+	symbolImage?: string;
+
+	showSettings: boolean;
+	showSettingsForType: boolean;
+	showInConstructionBox: boolean;
+
 	// translation id for predefined comps
 	description: string;
+
+	isRotatable?: boolean;
 	rotation: number;
-	category: 'basic' | 'plug' | 'io' | 'user';
+
+	numOutputs: number;
+
+	numInputs: number;
+	minInputs: number;
+	maxInputs: number;
+
+	hasPlugIndex?: boolean;
+	hasLabel?: boolean;
+
+	width: number;
+
+	options?: number[];
+	optionsConfig?: {
+		name: string,
+		min: number,
+		max: number
+	}[];
+	onOptionsChanged?: (element?: Element) => void;
+
+	edit?: (typeId: number, id: number, projectsSer: ProjectsService) => void;
+	canEditType?: boolean;
+
+	// only used for udcs, order must be like plugIndices
+	labels?: string[];
+	calcLabels?: (element?: Element) => string[];
 }
 
 export function isElementType(object: any): object is ElementType {
