@@ -6,6 +6,9 @@ import {WireGraphics} from './wire-graphics';
 import {ComponentGraphics} from './component-graphics';
 import {ButtonGraphics} from './button-graphics';
 import {LeverGraphics} from './lever-graphics';
+import {ElementTypeId} from '../../element-types/element-type-ids';
+import {TextGraphics} from './text-graphics';
+import {InputOutputGraphics} from './input-output-graphics';
 
 export abstract class LGraphicsResolver {
 
@@ -14,26 +17,36 @@ export abstract class LGraphicsResolver {
 	}
 
 	public static getLGraphicsFromElement(scale: number, element: Element, parentProjectIdentifier?: string): LGraphics {
-		if (element.typeId === 0) {
-			return new WireGraphics(scale, element);
-		} else if (this.elementProviderService.isButtonElement(element.typeId)) {
-			return new ButtonGraphics(scale, element, parentProjectIdentifier);
-		} else if (this.elementProviderService.isLeverElement(element.typeId)) {
-			return new LeverGraphics(scale, element, parentProjectIdentifier);
-		} else {
-			return new ComponentGraphics(scale, element);
+		switch (element.typeId) {
+			case ElementTypeId.WIRE:
+				return new WireGraphics(scale, element);
+			case ElementTypeId.BUTTON:
+				return new ButtonGraphics(scale, element, parentProjectIdentifier);
+			case ElementTypeId.LEVER:
+				return new LeverGraphics(scale, element, parentProjectIdentifier);
+			case ElementTypeId.TEXT:
+				return new TextGraphics(scale, element);
+			case ElementTypeId.INPUT:
+			case ElementTypeId.OUTPUT:
+				return new InputOutputGraphics(scale, element);
+			default:
+				return new ComponentGraphics(scale, element);
 		}
 	}
 
-	// wires are not supported !!
+	// wires and text are not supported !!
 	public static getLGraphicsFromType(scale: number, elemTypeId: number): LGraphics {
 		const elemType = this.elementProviderService.getElementById(elemTypeId);
-		if (this.elementProviderService.isButtonElement(elemTypeId)) {
-			return new ButtonGraphics(scale, elemType);
-		} else if (this.elementProviderService.isLeverElement(elemTypeId)) {
-			return new LeverGraphics(scale, elemType);
-		} else {
-			return new ComponentGraphics(scale, elemType);
+		switch (elemTypeId) {
+			case ElementTypeId.BUTTON:
+				return new ButtonGraphics(scale, elemType);
+			case ElementTypeId.LEVER:
+				return new LeverGraphics(scale, elemType);
+			case ElementTypeId.INPUT:
+			case ElementTypeId.OUTPUT:
+				return new InputOutputGraphics(scale, elemType);
+			default:
+				return new ComponentGraphics(scale, elemType);
 		}
 	}
 
