@@ -19,6 +19,9 @@ export class Grid {
 				graphics.drawRect(i * environment.gridPixelWidth, j * environment.gridPixelWidth, 1 / scale, 1 / scale);
 			}
 		}
+		// graphics.beginFill(0, 0);
+		// graphics.lineStyle(1 / scale, 0xffffff);
+		// graphics.drawRect(0, 0, environment.gridPixelWidth * environment.chunkSize, environment.gridPixelWidth * environment.chunkSize);
 		this._gridGeometries.set(scale, graphics.geometry);
 		return graphics.geometry;
 	}
@@ -27,15 +30,15 @@ export class Grid {
 		return new PIXI.Graphics(this.getGridGeometry(scale));
 	}
 
-	public static getGridPosForPixelPos(point: PIXI.Point): PIXI.Point {
+	public static getGridPosForPixelPos(point: PIXI.IPoint): PIXI.Point {
 		return new PIXI.Point(Math.floor(point.x / environment.gridPixelWidth), Math.floor(point.y / environment.gridPixelWidth));
 	}
 
-	public static getFloatGridPosForPixelPos(point: PIXI.Point): PIXI.Point {
+	public static getFloatGridPosForPixelPos(point: PIXI.IPoint): PIXI.Point {
 		return new PIXI.Point(point.x / environment.gridPixelWidth, point.y / environment.gridPixelWidth);
 	}
 
-	public static getPixelPosForGridPos(point: PIXI.Point): PIXI.Point {
+	public static getPixelPosForGridPos(point: PIXI.IPoint): PIXI.Point {
 		return new PIXI.Point(point.x * environment.gridPixelWidth, point.y * environment.gridPixelWidth);
 	}
 
@@ -46,25 +49,25 @@ export class Grid {
 		return Grid.getPixelPosForGridPos(newPoint);
 	}
 
-	public static getLocalChunkPixelPosForGridPosWireStart(point: PIXI.Point): PIXI.Point {
+	public static getLocalChunkPixelPosForGridPosWire(point: PIXI.Point): PIXI.Point {
 		const newPoint = point.clone();
 		newPoint.x += 0.5;
 		newPoint.y += 0.5;
 		return Grid.getLocalChunkPixelPosForGridPos(newPoint);
 	}
 
-	public static getPixelPosForGridPosWire(point: PIXI.Point): PIXI.Point {
+	public static getPixelPosForGridPosWire(point: PIXI.IPoint): PIXI.Point {
 		return Grid.getPixelPosForGridPos(new Point(point.x + 0.5, point.y + 0.5));
 	}
 
-	public static getPixelPosForPixelPosOnGridWire(point: PIXI.Point): PIXI.Point {
+	public static getPixelPosForPixelPosOnGridWire(point: PIXI.IPoint): PIXI.Point {
 		const gridPos = Grid.getGridPosForPixelPos(point);
 		gridPos.x += 0.5;
 		gridPos.y += 0.5;
 		return Grid.getPixelPosForGridPos(gridPos);
 	}
 
-	public static getPixelPosOnGridForPixelPos(point: PIXI.Point): PIXI.Point {
+	public static getPixelPosOnGridForPixelPos(point: PIXI.IPoint): PIXI.Point {
 		const gridPos = Grid.getGridPosForPixelPos(point);
 		return Grid.getPixelPosForGridPos(gridPos);
 	}
