@@ -2,6 +2,5 @@ export interface SimulationUnit {
 	typeId: number;
 	inputs: number[];
 	outputs: number[];
-	op1?: number;
-	op2?: number;
+	options: number[];
 }
